Added pre- and post-authentication filters
Logging filters split in to two so that we always add fish tags before requests reach the security filters. We can't add them after as not all requests will go through the security filter. For the post-authentication filter just add the authenticated users username to the ThreadContext for logging along with the request UUID generated at the pre-authentication stage