Display logged in user in nav bar and ability to log out
Uses Spring Security JSP tags to detect if a user is logged in and displays the username in the nav bar if so. Provides a logout button to log the current user out and invalidate the session.

If no user is logged in, then display controls on the nav bar that allows the user to log in. Although these are not currently functional in this commit.

In order to use Spring Forms, we need a model object to bind to, hence the currently slightly useless AppUserDetails object.