Allow admins to edit very basic user details. Strip back unimplemented code
First version of persistence allowing an admin to enable/disable a member.

Controllers that do nothing have been removed and simply registered in the servlet context configuration to map to the skeleton .jsp's

Also slightly reconfigured to view resolver to set the prefix to 'META-INF/jsp/'. Saves us having to specify 'jsp' at the root of the path if ever returning a view name