Significant simplification of the registration process
New users are instantly registered after passing the "bot" test. The user is then automatically logged in. The user chooses their own password at the sign-up dialog so a temporary one isn't generated and sent in an email. This saves the user from having to log in and change the password