Refactor of how user details are stored and updated
The User class was not a full implementation that retrieved all user details. This is now left to Springs' JdbcUserDetailsManager to load the complete record. User has been removed.

UserDetail contained redundant properties and was easily confused with Spring UserDetails interface, so has been renamed SupplementalUserDetails.

Maintain the separations between the authenticated principal and the additional user information, so that additional information isn't carried around with the principal. The other option is to combine the two and have a custom Entity class that replicates the User class, and embeds user info. This approach has not been taken.