Use Tomcat to load image resources from the filesystem
Allow's any environment (including local development) to be configured in a consistent way to load (image) assets not part of the .war. Previously handled by Apache httpd directives to not proxy requests for /images, loading them from the sites static document directory. This only worked for server deployments so was not portable.