Fix login issues where CSRF token expires
The forms on the home page request a CSRF token that is associated with a session as soon as the page loads. When this session expires the token is invalid. If the user tries to log in, the invalid session handler redirects the user back to the home page with a new token. This forces the user to log in twice, even though the initial attempt is valid.

The recommended technique (according to Spring) is to request a CSRF token only on form submission, which is a simple REST service call.