Refactored user management. Fix password change bug
Merged users and user_details tables to simplify the management and querying of user accounts and additional user details. Less code duplication and removal of superflous classes. More consistent interface. Still need to reinstate ability to change group membership.

Fixed a bug in the password change validator that essentially cached the details of the principal that first attempted to change their password, using that principal for all further password validation.