Added association between users and events
When an event is completed, all participating members are recorded. There is a many-to-many association between members and events they have participated in. This allows members to see which events they have completed, and to see which members participated in which events.

When an event is completed all members that were participating will also have their accounts debited. This is all handled in one transaction to ensure consistency.