Secure Service Interfaces. Improvements to sending registration mails
Created ROLE_RUN_AS_NEW_USER to designate methods that are run by anonymous users during registration but may need elevated privileges to run other service methods.

Running as the System user temporarily replaces the current user in the security context and restores the original user on completion of the task.

Mail sending tasks are run in a separate thread to speed the sign-up process. It is also demarcated as not supporting transactions so that failure to send mails doesn't cause successful registration to roll back

Registration mails are sent as part of the user creation process rahter than having to be called separately