clubbed-13 Run completion not recognising distance as a valid numeric in IE
Also affects some Safari versions.

Event .keyCode value likely different across browsers and in fact deprecated. Newer methods are also not fully supported in all browsers, so this approach has too many caveats to be reliable.

Removed keydown handler to allow all characters to be entered (also allows things like copy/paste etc). Validation is performed in keyup. If the input text is not numeric, the input element is highlighted as being in error and the Completion button is disabled. Server side validation checks the numeric for a valid range to prevent negative distances being entered.